validation/requirements.txt
azure-cognitiveservices-vision-customvision
azure-iot-device
requests